Fix kind load docker-image error
